qa: fix "no active mgr (MGR_DOWN)" in fs:nfs suite
RCA showed that it is not the NFS code that lead to the warning since
the warning occured before the test cases started to execute, later on
after some discussion with the venky and greg, it was found that there
were some clog changes made recently which leads to this warning being
added to the clog.

fs:nfs test suite makes use of just 1 MGR, so adding a second one for
backup should prevent getting MGR_DOWN warnings.

Fixes: https://tracker.ceph.com/issues/65265
Signed-off-by: Dhairya Parmar <dparmar@redhat.com>