mds/PurgeQueue: reword pq_executing to pq_executing_items
pq_executing is a perf counter that keeps track of the items (of type
PurgeItem) in purge queue. The item here can be a file or dir. The name of
this perf counter can get easily confused with another perf counter
pq_executing_ops which keeps track of the total ops (leaves count if it's a
dir or total objects if the purge item is a file) in flight (i.e. items being
expired/purged). The name pq_executing can be changed to something more
distinct like pq_executing_items.

Fixes: https://tracker.ceph.com/issues/68676
Signed-off-by: Dhairya Parmar <dparmar@redhat.com>